style : 검색바 세팅
